Python/requirements.txt
RubikTwoPhase
pygame-ce
